Style : "스킬 쿨타임을 float -> int32로 자료형 변경"
Data
- 스킬 쿨타임 자료형 변경 float -> int32

DataManager
- 스킬 쿨타임 자료형 변경에 따라 GetInt()로 읽어들임

SkillData
- 쿨타임에 맞게 ms단위로 쿨타임 기록

+++ GameServer/Data.h
@@ -73,7 +73,7 @@ struct st_SkillData
 {
 	int32 SkillDataId;
 	string SkillName;
-	float SkillCoolTime;
+	int32 SkillCoolTime;
 	float SkillCastingTime;
 	int SkillDistance;
 	string SkillThumbnailImagePath;
